# src/BackEnd/Api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "心之旅",
        "version": "1.0"
    },
    "paths": {
...
        "/api/admin/appt/list": {
            "post": {
                "tags": [
                    "总台管理系统-预约管理"
                ],
                "summary": "预约列表",
                "operationId": "App\\Http\\Controllers\\Admin\\AppointmentController::GetList",
                "parameters": [
                    {
                        "name": "token",
                        "in": "header",
                        "description": "token",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [],
                                "properties": {
                                    "storeId": {
                                        "description": "门店编号",
                                        "type": "string",
                                        "default": "13888888888"
                                    },
                                    "state": {
                                        "description": "状态",
                                        "type": "string",
                                        "default": ""
                                    },
                                    "userId": {
                                        "description": "人员编号",
                                        "type": "string",
                                        "default": ""
                                    },
                                    "wcId": {
                                        "description": "小程序用户编号",
                                        "type": "string",
                                        "default": ""
                                    },
                                    "startDate": {
                                        "description": "预约开始时间",
                                        "type": "string",
                                        "default": ""
                                    },
                                    "endDate": {
                                        "description": "预约结束时间",
                                        "type": "string",
                                        "default": ""
                                    },
                                    "pageSize": {
                                        "description": "条数",
                                        "type": "number",
                                        "default": "10"
                                    },
                                    "page": {
                                        "description": "页数",
                                        "type": "number",
                                        "default": "1"
                                    },
                                    "searchKey": {
                                        "description": "关键字",
                                        "type": "string",
                                        "default": ""
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "成功",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "description": "状态码",
                                            "type": "number",
                                            "example": "200"
                                        },
                                        "msg": {
                                            "description": "提示信息",
                                            "type": "string",
                                            "example": "获取列表成功!"
                                        },
                                        "data": {
                                            "type": "object",
                                            "example": ""
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "失败",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "description": "状态码",
                                            "type": "number",
                                            "example": "500"
                                        },
                                        "msg": {
                                            "description": "提示信息",
                                            "type": "string",
                                            "example": "获取列表失败!"
                                        },
                                        "data": {
                                            "type": "object",
                                            "example": "{'page':['\\u8bf7\\u8f93\\u5165\\u5355\\u4f4d\\u540d\\u79f0\\uff01']}"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/admin/combo/list": {
            "post": {
                "tags": [
                    "总台管理系统-套餐管理"
                ],
                "summary": "套餐列表",
                "operationId": "App\\Http\\Controllers\\Admin\\ComboController::GetList",
                "parameters": [
                    {
                        "name": "token",
                        "in": "header",
                        "description": "token",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [],
                                "properties": {
                                    "storeId": {
                                        "description": "门店编号",
                                        "type": "string",
                                        "default": ""
                                    },
                                    "varietyId": {
                                        "description": "宠物品种编号",
                                        "type": "string",
                                        "default": ""
                                    },
                                    "searchKey": {
                                        "description": "关键字",
                                        "type": "string",
                                        "default": ""
                                    },
                                    "comboType": {
                                        "description": "套餐类型：0-主套餐  1-增项套餐",
                                        "type": "string",
                                        "default": ""
                                    },
                                    "state": {
                                        "description": "上架状态 0-待上架 1-已上架 2-已下架",
                                        "type": "string",
                                        "default": ""
                                    },
                                    "pageSize": {
                                        "description": "条数",
                                        "type": "number",
                                        "default": "10"
                                    },
                                    "page": {
                                        "description": "页数",
                                        "type": "number",
                                        "default": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "成功",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "description": "状态码",
                                            "type": "number",
                                            "example": "200"
                                        },
                                        "msg": {
                                            "description": "提示信息",
                                            "type": "string",
                                            "example": "获取列表成功!"
                                        },
                                        "data": {
                                            "type": "object",
                                            "example": ""
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "失败",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "description": "状态码",
                                            "type": "number",
                                            "example": "500"
                                        },
                                        "msg": {
                                            "description": "提示信息",
                                            "type": "string",
                                            "example": "获取列表失败!"
                                        },
                                        "data": {
                                            "type": "object",
                                            "example": "{'page':['\\u8bf7\\u8f93\\u5165\\u5355\\u4f4d\\u540d\\u79f0\\uff01']}"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/admin/store/list": {
            "post": {
                "tags": [
                    "总台管理系统-门店管理"
                ],
                "summary": "门店列表",
                "operationId": "App\\Http\\Controllers\\Admin\\StoreController::GetList",
                "parameters": [
                    {
                        "name": "token",
                        "in": "header",
                        "description": "token",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "pageSize": {
                                        "description": "条数",
                                        "type": "number",
                                        "default": "10"
                                    },
                                    "page": {
                                        "description": "页数",
                                        "type": "number",
                                        "default": "1"
                                    },
                                    "searchKey": {
                                        "description": "关键字",
                                        "type": "string",
                                        "default": ""
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "成功",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "description": "状态码",
                                            "type": "number",
                                            "example": "200"
                                        },
                                        "msg": {
                                            "description": "提示信息",
                                            "type": "string",
                                            "example": "获取列表成功!"
                                        },
                                        "data": {
                                            "type": "object",
                                            "example": "{'status':200,'msg':'\\u83b7\\u53d6\\u5217\\u8868\\u6210\\u529f!','data':{'pages':{'total':1,'pageNo':'1'},'data':[{'id':'1','name':'\\u6328\\u6253\\u4e86','phone':'1','lng':'111','lat':'1','address':'1','businessHoursStart':'1','businessHoursEnd':'1','type':1,'state':0,'deleted_at':null,'created_at':null,'updated_at':null},{'id':'xxx','name':'\\u597d\\u5389\\u5bb3\\u54e6','phone':'1','lng':'2','lat':'2','address':'2','businessHoursStart':'2','businessHoursEnd':'2','type':1,'state':0,'deleted_at':null,'created_at':null,'updated_at':null},{'id':'\\u5fb7\\u74e6\\u8fbe','name':'ddx\\u738b\\u60f3\\u6328\\u6253\\u00b7','phone':'1','lng':'2','lat':'1','address':'11','businessHoursStart':'1','businessHoursEnd':'11','type':1,'state':0,'deleted_at':null,'created_at':null,'updated_at':null}]}}"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "失败",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "description": "状态码",
                                            "type": "number",
                                            "example": "500"
                                        },
                                        "msg": {
                                            "description": "提示信息",
                                            "type": "string",
                                            "example": "获取列表失败!"
                                        },
                                        "data": {
                                            "type": "object",
                                            "example": "{'page':['\\u8bf7\\u8f93\\u5165\\u5355\\u4f4d\\u540d\\u79f0\\uff01']}"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/admin/user/list": {
            "post": {
                "tags": [
                    "总台管理系统-人员管理"
                ],
                "summary": "人员列表",
                "operationId": "App\\Http\\Controllers\\Admin\\UserController::GetList",
                "parameters": [
                    {
                        "name": "token",
                        "in": "header",
                        "description": "token",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "storeId",
                                    "post"
                                ],
                                "properties": {
                                    "storeId": {
                                        "description": "门店编号",
                                        "type": "number",
                                        "default": "13888888888"
                                    },
                                    "post": {
                                        "description": "职位",
                                        "type": "string",
                                        "default": ""
                                    },
                                    "gender": {
                                        "description": "性别",
                                        "type": "string",
                                        "default": ""
                                    },
                                    "pageSize": {
                                        "description": "条数",
                                        "type": "number",
                                        "default": "10"
                                    },
                                    "page": {
                                        "description": "页数",
                                        "type": "number",
                                        "default": "1"
                                    },
                                    "searchKey": {
                                        "description": "关键字",
                                        "type": "string",
                                        "default": ""
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "成功",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "description": "状态码",
                                            "type": "number",
                                            "example": "200"
                                        },
                                        "msg": {
                                            "description": "提示信息",
                                            "type": "string",
                                            "example": "获取列表成功!"
                                        },
                                        "data": {
                                            "type": "object",
                                            "example": ""
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "失败",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "description": "状态码",
                                            "type": "number",
                                            "example": "500"
                                        },
                                        "msg": {
                                            "description": "提示信息",
                                            "type": "string",
                                            "example": "获取列表失败!"
                                        },
                                        "data": {
                                            "type": "object",
                                            "example": "{'page':['\\u8bf7\\u8f93\\u5165\\u5355\\u4f4d\\u540d\\u79f0\\uff01']}"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/admin/pet/list": {
            "get": {
                "tags": [
                    "总台管理系统-用户管理"
                ],
                "summary": "获取宠物列表",
                "operationId": "App\\Http\\Controllers\\Admin\\WechatUserController::GetPetList",
                "parameters": [
                    {
                        "name": "token",
                        "in": "header",
                        "description": "token",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "wcId",
                        "in": "query",
                        "description": "用户id",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "description": "状态码",
                                            "type": "number",
                                            "example": "200"
                                        },
                                        "msg": {
                                            "type": "string",
                                            "example": "获取宠物列表成功!"
                                        },
                                        "data": {
                                            "type": "object",
                                            "example": "[{'id':1,'avatar':'11','nickname':'11','breed':'11','gender':'11'}]"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "security": [
        []
    ]
}